refactor(bot): extract play command execute logic into handlers/ (Phase 2)
